fix: billion catalogue collapsed to 211 plans, and a silent pricing failure
The billion slug was built from (country, data, days, type), but its
daily plans carry no total quota and always report days=1, so dozens of
SKUs per country produced one slug and overwrote each other: 1866
products became 211 rows. Include the daily allowance and throttle, and
fall back to a SKU suffix when a slug is already taken by another SKU.

updateAllVndPrices compared a bare parameter to 0, so Postgres inferred
it as integer and rejected a real exchange rate. Its caller logs and
swallows the error, so every sync reported success while local plans
kept a stale usdPrice.

+++ src/plans/infrastructure/persistence/relational/repositories/plan.repository.ts
@@ -438,8 +438,13 @@ export class PlansRelationalRepository implements PlanRepository {
 
     // Local plans have no supplier dollar price at all, so derive one from the
     // VND price. Without this they reported their đồng figure as dollars.
+    // `$1::numeric` on both uses: with a bare `$1 > 0` Postgres infers the
+    // parameter as integer from the literal, so a real exchange rate
+    // (25887.097637) was rejected and this whole method threw. The caller
+    // logs and swallows that, so the failure was invisible while local plans
+    // silently kept a stale usdPrice.
     await this.plansRepository.query(
-      `UPDATE "plan" SET "usdPrice" = ROUND("vndPrice"::numeric / $1, 2) WHERE "deletedAt" IS NULL AND "isLocalInventory" = true AND $1 > 0`,
+      `UPDATE "plan" SET "usdPrice" = ROUND("vndPrice"::numeric / $1::numeric, 2) WHERE "deletedAt" IS NULL AND "isLocalInventory" = true AND $1::numeric > 0`,
       [rate],
     );
   }
